--- mypy.ini
@@ -1,16 +1,0 @@
-[mypy]
-files = src/freva_stats_service
-strict = True
-
-warn_unused_ignores = True
-warn_unreachable = True
-show_error_codes = True
-install_types = True
-non_interactive = True
-warn_unused_configs = True
-disallow_untyped_defs = True
-disallow_incomplete_defs = True
-check_untyped_defs = True
-disallow_untyped_decorators = True
-warn_redundant_casts = True
-ignore_missing_imports = True
