+++ src/freva_storage_service/stats/databrowser.py
